PR-B-2: extract MemberLookup (binder-facing facade over ClrTypeUtilities)
Per the Binder decomposition plan. MemberLookup is the pure
"given a type T and a name N, return the candidates" facade. It is
constructor-injected with BinderContext (no back-reference to Binder)
and delegates raw CLR member walks to ClrTypeUtilities.

Moved into src/Core/CodeAnalysis/Binding/MemberLookup.cs:
- HasMatchingMethodForClrSignature (static)
- FindMatchingProperty (static)
- TryGetPropertyIncludingInherited (static)
- TryGetAsyncEnumerableElementType (static)
- EnumerateSelfAndInterfaces (static)
- TryGetClrDictionaryTypes (static)
- TryGetClrEnumerableElementType (static)
- TryGetClrPatternEnumerableElementType (static)
- TryGetClrCurrentMemberType (static)
- TryGetUserPatternEnumerableElementType (static)
- CollectClrParameterNames (static)
- CollectClrConstructorParameterNames (static)
- TryGetDelegateFunctionType (static)
- TryGetNullableConstructedType (instance, uses BinderContext.References)
- TryResolveClrIndexer (refactored: takes pre-bound
  ImmutableArray<BoundExpression> instead of ExpressionSyntax,
  so the facade itself does no BoundExpression construction;
  callers bind once before invoking the lookup, preserving IL)
- CollectImportedExtensionMethods (instance, walks References)
- GetImportedExtensionClasses (instance, populates BinderContext caches)
- IsStaticClass (static)
- HasExtensionAttribute (static)

Deliberately left on Binder:
- The BindAccessor* / BindCallExpression / BindIndexedRead / Bind*
  family — they construct BoundExpression and/or emit diagnostics
  (call-site shaping). Migrate to B-9 ExpressionBinder.
- BindClrMethodGroupConversion and other BindClr*Conversion — those
  are conversion classification, not lookup. Migrate to B-3
  ConversionClassifier.
- TryReorder* / TryReport* overload-resolution glue — emits
  diagnostics; migrate to B-4 OverloadResolver.
- TryBuildDisposeCall / TryApplyUserDefinedImplicitArgumentConversion
  — constructs BoundExpression. Migrate to B-3 / B-9.
- TryResolve* name-resolution helpers (e.g. TryResolveScope,
  TryResolveUserMember) — different facet from member lookup
  (binds symbols via scopes, not types).
- The static BindGlobalScope path and TryInferExpressionType — they
  would require plumbing a MemberLookup instance into static
  contexts; deferred per the plan.

Wiring: Binder constructs a single private readonly MemberLookup
memberLookup = new MemberLookup(binderCtx) and routes call sites
to it. No public/internal API visible to Compilation.cs,
HoverComputer.cs, or any test was renamed.

Validation:
- Build clean (0 warnings, 0 errors).
- Core.Tests: 2120 passed / 1 skipped — matches main baseline.
- RefactoringBaselineTests: 92 passed / 1 skipped (the
  RegenerateBaseline gate) — zero IL hash diffs.
- Interpreter.Tests / LanguageServer.Tests / Sdk.Tests: all green.
- Binder.cs: 18,546 → 18,027 LoC (-519).
- MemberLookup.cs: 736 LoC.

Next: PR-B-3 ConversionClassifier.
